Make fsck return more info, also add dry-run mode.

+++ src/server/pfs/server/driver.go
@@ -1561,17 +1561,18 @@ func (d *driver) deleteAll(ctx context.Context) error {
 	return nil
 }
 
-func (d *driver) fsck(ctx context.Context) error {
+func (d *driver) fsck(ctx context.Context, dryRun bool) (*pfs.FsckResponse, error) {
 	objClient, err := d.getObjectClient()
 	if err != nil {
-		return err
+		return nil, err
 	}
 	l, err := objClient.ListObjects(ctx, &pfs.ListObjectsRequest{})
 	if err != nil {
-		return err
+		return nil, err
 	}
 	var eg errgroup.Group
 	deleteObjectsRequest := &pfs.DeleteObjectsRequest{}
+	result := &pfs.FsckResponse{}
 	var lock sync.Mutex
 	for {
 		object, err := l.Recv()
@@ -1579,7 +1580,7 @@ func (d *driver) fsck(ctx context.Context) error {
 			if err == io.EOF {
 				break
 			}
-			return err
+			return nil, err
 		}
 		eg.Go(func() error {
 			resp, err := objClient.CheckObject(ctx, &pfs.CheckObjectRequest{object})
@@ -1589,26 +1590,31 @@ func (d *driver) fsck(ctx context.Context) error {
 			if !resp.Exists {
 				lock.Lock()
 				defer lock.Unlock()
-				deleteObjectsRequest.Objects = append(deleteObjectsRequest.Objects, object)
-				if len(deleteObjectsRequest.Objects) > 100 {
-					request := deleteObjectsRequest
-					deleteObjectsRequest = &pfs.DeleteObjectsRequest{}
-					eg.Go(func() error {
-						_, err := objClient.DeleteObjects(ctx, request)
-						return err
-					})
+				result.CorruptedObjects = append(result.CorruptedObjects, object)
+				if !dryRun {
+					deleteObjectsRequest.Objects = append(deleteObjectsRequest.Objects, object)
+					if len(deleteObjectsRequest.Objects) > 100 {
+						request := deleteObjectsRequest
+						deleteObjectsRequest = &pfs.DeleteObjectsRequest{}
+						eg.Go(func() error {
+							_, err := objClient.DeleteObjects(ctx, request)
+							return err
+						})
+					}
 				}
 			}
 			return nil
 		})
 	}
 	if err := eg.Wait(); err != nil {
-		return err
+		return nil, err
 	}
-	if _, err := objClient.DeleteObjects(ctx, deleteObjectsRequest); err != nil {
-		return err
+	if !dryRun {
+		if _, err := objClient.DeleteObjects(ctx, deleteObjectsRequest); err != nil {
+			return nil, err
+		}
 	}
-	return nil
+	return result, nil
 }
 
 func (d *driver) applyWrites(resp *etcd.GetResponse, tree hashtree.OpenHashTree) error {
